Fix: Integrate summarization with existing background.js instead of creating a new file
