+++ extension/vite.config.ts
@@ -104,7 +104,7 @@ export default defineConfig({
     rollupOptions: {
       input: {
         popup: resolve(__dirname, 'src/popup.tsx'),
-        background: resolve(__dirname, 'src/background.ts'),
+        background: resolve(__dirname, 'background.js'),
         settings: resolve(__dirname, 'src/settings/index.ts'),
         history: resolve(__dirname, 'src/history.tsx'),
         content: resolve(__dirname, 'src/content.ts')
